Replaced step_four.sh with original

+++ step_four.sh
@@ -1,21 +1,29 @@
 #!/bin/bash
 
 
-find . -name "pom.xml" -type f | while read -r POM_FILE; do
-
-    cp "$POM_FILE" "${POM_FILE}.bak"
-
-    sed -i 's|<url>http://|<url>https://|g' "$POM_FILE"
-    sed -i 's|<repositoryUrl>http://|<repositoryUrl>https://|g' "$POM_FILE"
-    sed -i 's|<connection>http://|<connection>https://|g' "$POM_FILE"
-    sed -i 's|<developerConnection>http://|<developerConnection>https://|g' "$POM_FILE"
-    sed -i 's|http://repo1.maven.org/maven2|https://repo1.maven.org/maven2|g' "$POM_FILE"
-    sed -i 's|http://|https://|g' "$POM_FILE" 
-done
+POM_FILE="pom.xml"
 
 
-echo "All applicable http:// URLs in pom.xml files have been replaced with https://"
+if [ ! -f "$POM_FILE" ]; then
+    echo "No pom.xml file found in current directory."
+    exit 1
+fi
 
+
+# Backup original pom.xml
+cp "$POM_FILE" "${POM_FILE}.bak"
+
+
+# Replace all http:// with https:// in the context of dependencies and repositories
+sed -i 's|<url>http://|<url>https://|g' "$POM_FILE"
+sed -i 's|<repositoryUrl>http://|<repositoryUrl>https://|g' "$POM_FILE"
+sed -i 's|<connection>http://|<connection>https://|g' "$POM_FILE"
+sed -i 's|<developerConnection>http://|<developerConnection>https://|g' "$POM_FILE"
+sed -i 's|<distributionManagement>.*http://|<distributionManagement>https://|g' "$POM_FILE"
+sed -i 's|http://repo1.maven.org/maven2|https://repo1.maven.org/maven2|g' "$POM_FILE"
+sed -i 's|http://|https://|g' "$POM_FILE"  # fallback, more aggressive (can be commented if too broad)
+
+
+echo "All applicable http:// URLs in $POM_FILE have been replaced with https://"
 
 exit 0
-
